Implamented Firing left, Started Universal App
Have started developing the app as a universal application.
I want it to be able to play over an apple tv with an iphone
being used as a joy pad.

Have created and implemented the firing left animation. just need
to rotate the magice missile so it faces the correct direction.