Begin Map Generation
Began the implementation of maps. This uses a bases maze generation
algorithm as a base. It should also be quite easy to subclass or
modify to create new generators which create different style levels

Also updated the README file to support gits markup language. Will
go over it and add more detail at some point but things are changing
so quickly at the moment there is little point.